+++ src/ui-input/ui-input.js
@@ -172,7 +172,7 @@ export class UIInput extends HTMLElement {
                     font-size: 0.9rem;
                     font-family: var(--ui-fontFamily);
                     font-variation-settings: var(--ui-input-fontVariation);
-                    accent-color: var(--ui-primary-bg);
+                    accent-color: var(--ui-primary);
                     background-color: transparent !important;
                 }
 
@@ -185,11 +185,11 @@ export class UIInput extends HTMLElement {
                 }
 
                 .container:has(input:focus) {
-                    border-color: var(--ui-primary-bg);
+                    border-color: var(--ui-primary);
                 }
 
                 :host([invalid]) .container {
-                    border-color: var(--ui-destructive-bg);
+                    border-color: var(--ui-destructive);
                 }
 
                 ::slotted([slot="title"]) {